update antra with json
